pyproject.toml
[tool.poetry]
name = "allergia-bot"
version = "0.1.0"
description = ""
authors = ["twoballs-ai <twoballs20@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
aiogram = "^3.12.0"
aiohttp = "^3.10.5"
python-dotenv = "^1.0.1"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
